Cleaned up testing infrastructure.
