--- tests/__init__.py
@@ -1,1 +1,0 @@
-from glinkfix import tools  # noqa
